@@ -1,21 +1,14 @@
-// The Vue build version to load with the `import` command
-// (runtime-only or standalone) has been set in webpack.base.conf with an alias.
-import Vue from 'vue'
-import App from './App'
+import { createApp } from 'vue'
+
+import App from './App.vue'
 import router from './router'
+import VueClickAway from "vue3-click-away";
+import { createVuestic } from 'vuestic-ui'
+import config from '../vuestic.config.js'
+import './style.css'
 
-import VueCookies from 'vue-cookies'
-import Buefy from 'buefy'
-import 'buefy/dist/buefy.css'
-
-Vue.use(Buefy)
-Vue.use(VueCookies)
-Vue.config.productionTip = false
-
-/* eslint-disable no-new */
-new Vue({
-  el: '#app',
-  router,
-  components: { App },
-  template: '<App/>'
-})
+const app = createApp(App)
+app.use(router)
+app.use(VueClickAway);
+app.use(createVuestic({config}));
+app.mount('#app')
