@@ -4,8 +4,8 @@ package com.trevorism.gcloud
  * @author tbrooks
  */
 
-this.metaClass.mixin(cucumber.api.groovy.Hooks)
-this.metaClass.mixin(cucumber.api.groovy.EN)
+this.metaClass.mixin(io.cucumber.groovy.Hooks)
+this.metaClass.mixin(io.cucumber.groovy.EN)
 
 def contextRootContent
 def pingContent
