@@ -1,19 +1,13 @@
+<script setup>
+import { RouterView } from 'vue-router'
+</script>
+
 <template>
-  <div id="app">
+  <div>
     <router-view></router-view>
   </div>
 </template>
 
-<script>
-
-export default {
-  name: 'App'
-}
-</script>
-
 <style>
-#app {
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
+
 </style>
