@@ -1,12 +1,12 @@
 <template>
-  <div id="admin">
-    <HeaderBar></HeaderBar>
+  <div class="grid justify-items-center" id="admin">
+    <HeaderBar :local="true"></HeaderBar>
     <div id="adminText">Only accessible to logged-in Admin users</div>
   </div>
 </template>
 
 <script>
-import HeaderBar from './HeaderBar'
+import HeaderBar from '@trevorism/ui-header-bar'
 
 export default {
   name: 'Admin',
