@@ -1,57 +1,63 @@
 <template>
-  <div id="change">
-    <HeaderBar></HeaderBar>
-    <section class="container formWidth">
-      <div class="is-size-4 mt-6 mb-6">Change Password on Trevorism</div>
-      <form class="loginBorder">
+  <div class="grid justify-items-center" id="change">
+    <HeaderBar :local="true"></HeaderBar>
+    <div class="container grid justify-items-center">
+      <h2 class=" text-xl font-bold py-6 my-6">Change Password on Trevorism</h2>
+
+      <va-form class="border-2 rounded-md w-80">
         <div class="mx-4 mt-4 mb-4">
-          <b-field label="Username">
-            <b-input type="text"
-                     required
-                     validation-message="Must be at least 3 characters"
-                     minlength="3"
-                     v-model="username"
-                     :autofocus="true">
-            </b-input>
-          </b-field>
-          <b-field label="Current Password">
-            <b-input type="password"
-                     required
-                     validation-message="Must be at least 6 characters"
-                     minlength="6"
-                     v-model="currentPassword"></b-input>
-          </b-field>
-          <b-field label="New Password">
-            <b-input type="password"
-                     required
-                     validation-message="Must be at least 6 characters"
-                     minlength="6"
-                     v-model="newPassword">
-            </b-input>
-          </b-field>
-          <b-field label="Repeat New Password">
-            <b-input type="password"
-                     required
-                     validation-message="Must be at least 6 characters and match the new password"
-                     minlength="6"
-                     v-model="repeatPassword">
-            </b-input>
-          </b-field>
-          <div class="is-centered has-text-centered">
-            <button class="button is-primary" :disabled="disabled" @click="invokeButton">
+          <va-input v-model="username"
+                    type="text"
+                    class="mb-6 w-full"
+                    :rules="[(v) => v.length >= 3]"
+                    required
+                    error-messages="Must be at least 3 characters"
+                    autofocus="true"
+                    label="Username"
+                    minlength="3">
+          </va-input>
+          <va-input type="password"
+                    class="mb-6 w-full"
+                    required
+                    :rules="[(v) => v.length >= 6]"
+                    error-messages="Must be at least 6 characters"
+                    label="Current Password"
+                    minlength="6"
+                    v-model="currentPassword">
+          </va-input>
+          <va-input type="password"
+                    class="mb-6 w-full"
+                    required
+                    :rules="[(v) => v.length >= 6]"
+                    error-messages="Must be at least 6 characters"
+                    label="Password"
+                    minlength="6"
+                    v-model="newPassword">
+          </va-input>
+          <va-input type="password"
+                    class="mb-6 w-full"
+                    required
+                    :rules="[(v) => v.length >= 6, (v) => (v === this.newPassword)]"
+                    error-messages="Must be at least 6 characters and match the new password"
+                    label="Repeat Password"
+                    minlength="6"
+                    v-model="repeatPassword">
+          </va-input>
+          <div class="grid justify-items-center">
+            <va-button color="success" :disabled="disabled" @click="invokeButton">
               Submit
-            </button>
-            <b-loading :is-full-page="false" :active.sync="disabled" :can-cancel="false"></b-loading>
+            </va-button>
           </div>
         </div>
-      </form>
-      <div class="has-text-centered has-text-danger">{{errorMessage}}</div>
-    </section>
+      </va-form>
+      <va-alert v-if="errorMessage.length > 0" class="w-80 text-center" color="danger">{{errorMessage}}</va-alert>
+
+    </div>
   </div>
 </template>
 
 <script>
-import HeaderBar from './HeaderBar'
+import HeaderBar from '@trevorism/ui-header-bar'
 import axios from 'axios'
 
 export default {
@@ -104,11 +110,5 @@ export default {
 </script>
 
 <style scoped>
-  .loginBorder {
-    border: #dddddd 1px solid;
-    background: #efefef;
-  }
-  .formWidth {
-    width: 400px
-  }
+
 </style>
