@@ -1,78 +1,83 @@
 <template>
-  <div id="improvement">
-    <HeaderBar></HeaderBar>
-    <div class="container">
-      <div class="title mt-6">
+  <div class="grid justify-items-center" id="improvement">
+    <HeaderBar :local="true"></HeaderBar>
+    <div class="container px-4">
+      <h1 class="text-3xl font-bold -ml-4 mb-2">
         Scientific Improvement
-      </div>
+      </h1>
 
-      <div class="content">
+      <p class="mb-2">
+        December 2020
+      </p>
+
+
+      <p class="mb-2">
           Improvement is the key to success. Society wants and therefore rewards productivity,
           so gains in this area are very valuable. I will describe a framework for sustained
           improvement that can be applied to oneself, and can be scaled to teams and organizations.
-          The secret is in computer science and software engineering! Before that secret is revealed,
+          The secret lies within computer science and software engineering! Before that secret is revealed,
           let’s introduce a few concepts..
-      </div>
+      </p>
 
-      <section class="content">
-        <h4>Scientific Method</h4>
-      </section>
+      <h2 class="text-xl font-bold -ml-4 mb-2">
+        Scientific Method
+      </h2>
 
-      <div class="content">
+      <p class="mb-2">
         The Scientific Method is the best way to form new knowledge about the world.
         The feature that makes this process so effective is the rigor it imposes on knowledge creation;
         we must get repeatable, peer-reviewed, empirical results before hypotheses are accepted as facts.
         Outcomes are documented and the resultant facts offer themselves up to be disproven at any time.
-      </div>
+      </p>
 
-      <section class="content">
-        <h4>Computer Science</h4>
-      </section>
+      <h2 class="text-xl font-bold -ml-4 mb-2">
+        Computer Science
+      </h2>
 
-      <div class="content">
+      <p class="mb-2">
         Most think of Computer Science as the study of computation, but fundamentally computation is a process
         of taking inputs into a desired outcome. In our case, we want to take our life experiences as input,
         and make productivity gains the result. Can such a software system be built?
-      </div>
+      </p>
 
-      <section class="content">
-        <h4>Feedback Loop</h4>
-      </section>
+      <h2 class="text-xl font-bold -ml-4 mb-2">
+        Feedback Loop
+      </h2>
 
-      <div class="content">
+      <p class="mb-2">
         Achieving a long term goal must happen one step at a time. A feedback loop is an incredibly useful concept
         for ensuring each step brings you closer to your goal. It consists of three parts: a prediction;
         actions to bring about a result; and evaluation of the outcome. Evaluation is the hardest part --
         it takes experience to understand how to properly evaluate a process. The world is a dynamic place
         and external factors (luck) can have dramatic effects. Setting up objective measurements and appealing
-        to others are the best ways to evaluate in lieu of personal experience.
-      </div>
+        to others are the best ways to evaluate an outcome in lieu of personal experience.
+      </p>
 
-      <section class="content">
-        <h4>Engineering</h4>
-      </section>
+      <h2 class="text-xl font-bold -ml-4 mb-2">
+        Engineering
+      </h2>
 
-      <div class="content">
+      <p class="mb-2">
         So far, we’ve introduced concepts, but Engineering forces us to make these concepts tangible.
         We must consider constraints on time and energy and make practical decisions. Setting quantitative expectations,
         and measuring the results are important facets of engineering, but nothing compares to taking immediate action.
         Note that thought experiments are the opposite of engineering gains in productivity.
-        We weigh the value against the cost of anything that uses resources.
-      </div>
+        We weigh the potential value of any action against the cost of valuable resources.
+      </p>
 
-      <section class="content">
-        <h4>Scientific Improvement</h4>
-      </section>
+      <h2 class="text-xl font-bold -ml-4 mb-2">
+        Scientific Improvement
+      </h2>
 
-      <div class="content">
+      <p class="mb-2">
         With the above concepts as tools, we can now put together a framework for scientific improvement.
-        First, set up several feedback loops while working on a long term goal. Remember it has three parts:
+        First, one should establish several feedback loops as one progresses towards a long term goal. Remember it has three parts:
         a prediction (hypothesis), action (an experiment) and evaluation. Create visibility by writing down
         and publicizing the feedback loops. Here are some important details:
-      </div>
+      </p>
 
-      <section class="content">
-        <ol>
+      <div class="mb-2 ml-2">
+        <ul class="list-disc ml-6">
           <li>The prediction should include what will happen in the experiment, and when the evaluation will be completed. </li>
           <li>
             The action should occur as soon as possible. From engineering, we know we have limited time and resources,
@@ -82,52 +87,48 @@
             The evaluation should include some quantitative measure of the result, or leverage the opinion of the best
             people you can find.
           </li>
-        </ol>
-      </section>
-
-      <div class="content">
-        The framework for scientific improvement is to create a feedback loop on your set of feedback loops.
-        We rigorously measure the effectiveness of the loops, attempting to:
+        </ul>
       </div>
 
-      <section class="content">
-        <ul>
+      <p class="mb-2">
+        The framework for scientific improvement is to create a feedback loop on your set of feedback loops.
+        We rigorously measure the effectiveness of the loops, attempting to:
+      </p>
+
+      <div class="mb-2 ml-2">
+        <ul class="list-disc ml-6">
           <li>Increase the number of feedback loops, using mistakes or gaps in the process as an opportunity to install another loop. </li>
           <li>Increase the frequency of their use. </li>
           <li>Shorten the length of time spent within the feedback loop.</li>
           <li>Improve the evaluation by attempting experiments over time, and by seeking out more expertise.</li>
         </ul>
-      </section>
+      </div>
 
-      <div class="content">
+      <p class="mb-2">
         Commit to using each feedback loop within a period of time, usually a week or a month. At the end of this time period,
         any unused feedback loops should be re-evaluated, either recommitted or removed. The recursive nature of this is an elegant way
         to automatically reinforce improvement.
-      </div>
+      </p>
 
-      <div class="content">
+      <p class="mb-2">
         We can use software to automate and visualize many aspects of this framework. Trevorism.com offers the tools to
         perform keep lists, set expectations, measure results, and visualize outcomes, all open-source and freely available.
         Through automation, one can achieve inhuman levels of productivity, as software organizes and aids you on your journey.
         From a competitive standpoint, it gives you an advantage over those who do not take control of their destiny.
-      </div>
+      </p>
 
-      <div class="content">
+      <p class="mb-2">
         Over time, you should see dramatic improvements in your productivity. By establishing a rigorous framework,
         you now have a provable documented way to improve. Since this is done scientifically, you can extend this to your
-        teams and organizations. It works at any level of abstraction.
-      </div>
+        teams and organizations. This system works for individuals, teams, and organizations. It is the key to success.
+      </p>
 
     </div>
-
-    <footer>
-      <hr/>
-    </footer>
   </div>
 </template>
 
 <script>
-import HeaderBar from '../HeaderBar'
+import HeaderBar from '@trevorism/ui-header-bar'
 
 export default {
   name: 'improvement',
