@@ -2,57 +2,73 @@ package com.trevorism.gcloud.webapi.controller
 
 import com.trevorism.http.async.AsyncHttpClient
 import com.trevorism.http.async.AsyncJsonHttpClient
-import io.swagger.annotations.Api
-import io.swagger.annotations.ApiOperation
-import io.swagger.annotations.Contact
-import io.swagger.annotations.Info
-import io.swagger.annotations.SwaggerDefinition
+import io.micronaut.http.HttpResponse
+import io.micronaut.http.MediaType
+import io.micronaut.http.annotation.Controller
+import io.micronaut.http.annotation.Get
+import io.swagger.v3.oas.annotations.Operation
+import io.swagger.v3.oas.annotations.media.Content
+import io.swagger.v3.oas.annotations.media.Schema
+import io.swagger.v3.oas.annotations.responses.ApiResponse
+import io.swagger.v3.oas.annotations.tags.Tag
+import org.apache.hc.core5.concurrent.FutureCallback
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
 
-import javax.ws.rs.GET
-import javax.ws.rs.Path
-import javax.ws.rs.Produces
-import javax.ws.rs.core.MediaType
-import javax.ws.rs.core.Response
-
-@Api("Root Operations")
-@SwaggerDefinition(
-        info = @Info(
-                description = "API",
-                version = "1",
-                title = "API",
-                contact = @Contact(name = "Trevor Brooks", url = "https://www.trevorism.com")
-        )
-)
-@Path("/")
+@Controller("/api")
 class RootController {
 
-    @ApiOperation(value = "Returns 'pong' if the application is alive")
-    @GET
-    @Path("ping")
-    @Produces(MediaType.APPLICATION_JSON)
+    private static final Logger log = LoggerFactory.getLogger(RootController)
+
+    @Tag(name = "Root Operations")
+    @Operation(summary = "Context Root of the Application")
+    @ApiResponse(
+            responseCode = "200", content = @Content(mediaType = "text/html", schema = @Schema(type = "string"))
+    )
+
+    @Tag(name = "Root Operations")
+    @Get(produces = MediaType.TEXT_HTML)
+    HttpResponse<List<String>> index() {
+        log.info("Hit context root")
+        HttpResponse.ok(['<a href="/api/ping">/ping</a>', '<a href="/api/help">/help</a>', '<a href="/api/version">/version</a>'])
+    }
+
+    @Tag(name = "Root Operations")
+    @Operation(summary = "Returns 'pong' on success")
+    @ApiResponse(
+            responseCode = "200", content = @Content(mediaType = "text/plain", schema = @Schema(type = "string"))
+    )
+    @Get(value = "/ping", produces = MediaType.TEXT_PLAIN)
     String ping() {
-        "pong"
+        return "pong"
     }
 
-    @ApiOperation(value = "Context root of the application")
-    @GET
-    String displayHelpLink(){
-        '<h1>API</h1><br/>Visit the help page at <a href="/api/help">/api/help'
+    @Tag(name = "Root Operations")
+    @Operation(summary = "This help page")
+    @ApiResponse(responseCode = "302")
+    @Get(value = "/help")
+    HttpResponse<String> help() {
+        return HttpResponse.redirect(new URI("../swagger-ui/index.html"))
     }
 
-    @ApiOperation(value = "Shows this help page")
-    @GET
-    @Path("help")
-    Response help(){
-        Response.temporaryRedirect(new URI("/swagger/index.html")).build()
+    @Tag(name = "Root Operations")
+    @Operation(summary = "Returns the version of the API")
+    @ApiResponse(
+            responseCode = "200",
+            content = @Content(mediaType = "text/plain",
+                    schema = @Schema(type = "string"))
+    )
+    @Get(value = "/version", produces = MediaType.TEXT_PLAIN)
+    String version() {
+        return "0-0-2"
     }
 
-    @ApiOperation(value = "Warms up the authorization service")
-    @GET
-    @Path("authWarmup")
+    @Tag(name = "Root Operations")
+    @Operation(summary = "Warms up the authorization service")
+    @Get(value = "/authWarmup")
     void warmupAuthService(){
         AsyncHttpClient client = new AsyncJsonHttpClient()
-        client.get("https://datastore.trevorism.com/ping")
-        client.get("https://auth.trevorism.com/ping").get()
+        client.get("https://datastore.trevorism.com/ping", {} as FutureCallback)
+        client.get("https://auth.trevorism.com/ping", {} as FutureCallback)
     }
 }
