@@ -6,22 +6,24 @@ import com.trevorism.data.Repository
 import com.trevorism.data.model.filtering.FilterBuilder
 import com.trevorism.data.model.filtering.SimpleFilter
 import com.trevorism.gcloud.webapi.model.*
-import com.trevorism.http.headers.HeadersJsonHttpClient
-import com.trevorism.http.util.ResponseUtils
+import com.trevorism.http.HttpClient
+import com.trevorism.http.JsonHttpClient
 import com.trevorism.https.SecureHttpClient
-import com.trevorism.secure.ClaimProperties
-import com.trevorism.secure.ClaimsProvider
-import com.trevorism.secure.ClasspathBasedPropertiesProvider
-import com.trevorism.secure.PropertiesProvider
+import com.trevorism.ClaimProperties
+import com.trevorism.ClaimsProvider
+import com.trevorism.ClasspathBasedPropertiesProvider
+import com.trevorism.PropertiesProvider
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
 
 import java.time.LocalDateTime
-import java.util.logging.Logger
+
 
 class DefaultUserSessionService implements UserSessionService {
 
-    private static final Logger log = Logger.getLogger(DefaultUserSessionService.class.getName())
+    private static final Logger log = LoggerFactory.getLogger(DefaultUserSessionService.class.getName())
 
-    private HeadersJsonHttpClient httpClient = new HeadersJsonHttpClient()
+    private HttpClient httpClient = new JsonHttpClient()
     private SecureHttpClient secureHttpClient = SecureHttpClientSingleton.getInstance().getSecureHttpClient()
     private Repository<ForgotPasswordLink> forgotPasswordLinkRepository = new FastDatastoreRepository<>(ForgotPasswordLink.class, secureHttpClient)
     private Repository<User> repository = new FastDatastoreRepository<>(User, secureHttpClient)
@@ -35,8 +37,7 @@ class DefaultUserSessionService implements UserSessionService {
         try {
             return invokeTokenRequest(json)
         } catch (Exception e) {
-            log.fine("Invalid login")
-            log.finer(e.message)
+            log.debug("Invalid login", e)
         }
         return null
     }
@@ -46,10 +47,10 @@ class DefaultUserSessionService implements UserSessionService {
         try {
             ClaimProperties claimProperties = ClaimsProvider.getClaims(token, propertiesProvider.getProperty("signingKey"))
             def response = httpClient.get("https://auth.trevorism.com/user/${claimProperties.id}", ["Authorization": "bearer ${token}".toString()])
-            User user = gson.fromJson(ResponseUtils.getEntity(response), User)
+            User user = gson.fromJson(response.value, User)
             return user
         } catch (Exception e) {
-            log.warning("Unable to find user: ${e.message}")
+            log.warn("Unable to find user", e)
             return User.NULL_USER
         }
     }
@@ -61,7 +62,7 @@ class DefaultUserSessionService implements UserSessionService {
         }
         String json = gson.toJson(registrationRequest)
         def response = httpClient.post("https://auth.trevorism.com/user", json, [:])
-        User user = gson.fromJson(ResponseUtils.getEntity(response), User)
+        User user = gson.fromJson(response.value, User)
         return !User.isNullUser(user)
     }
 
@@ -113,7 +114,7 @@ class DefaultUserSessionService implements UserSessionService {
     boolean changePassword(ChangePasswordRequest changePasswordRequest, String token) {
         String json = gson.toJson(changePasswordRequest)
         def response = httpClient.post("https://auth.trevorism.com/user/change", json, ["Authorization": "bearer ${token}".toString()])
-        String value = ResponseUtils.getEntity(response)
+        String value = response.value
         return value == "true"
     }
 
@@ -143,10 +144,10 @@ class DefaultUserSessionService implements UserSessionService {
     }
 
     private String invokeTokenRequest(String json) {
-        String result = ResponseUtils.getEntity(httpClient.post("https://auth.trevorism.com/token", json, [:]))
+        String result = httpClient.post("https://auth.trevorism.com/token", json, [:]).value
         if (result.startsWith("<html>"))
             throw new RuntimeException("Bad Request to get token")
-        log.fine("Successful login, token: ${result}")
+        log.trace("Successful login, token: ${result}")
         return result
     }
 }
