@@ -1,6 +1,7 @@
-package com.trevorism.gcloud.webapi.controller
+package com.trevorism.controller
 
-import org.junit.Test
+import com.trevorism.gcloud.webapi.controller.RootController
+import org.junit.jupiter.api.Test
 
 /**
  * @author tbrooks
@@ -10,18 +11,12 @@ class RootControllerTest {
     @Test
     void testRootControllerEndpoints(){
         RootController rootController = new RootController()
-        assert rootController.displayHelpLink().contains("/help")
+        assert rootController.index().getBody().get().join(";").contains("/help")
     }
 
     @Test
     void testRootControllerPing(){
         RootController rootController = new RootController()
         assert rootController.ping() == "pong"
-    }
-
-    @Test
-    void testRootControllerHelp(){
-        RootController rootController = new RootController()
-        assert rootController.help()
     }
 }
