@@ -1,13 +1,13 @@
 package com.trevorism.gcloud.webapi.service
 
 import com.trevorism.data.Repository
-import com.trevorism.data.model.filtering.SimpleFilter
 import com.trevorism.gcloud.webapi.model.*
-import com.trevorism.http.headers.HeadersJsonHttpClient
+import com.trevorism.http.HeadersHttpResponse
+import com.trevorism.http.HttpClient
 import com.trevorism.https.SecureHttpClient
-import org.apache.http.client.methods.CloseableHttpResponse
-import org.apache.http.entity.StringEntity
-import org.junit.Test
+import org.junit.jupiter.api.Test
+
+import static org.junit.jupiter.api.Assertions.assertThrows
 
 class DefaultUserSessionServiceTest {
 
@@ -25,17 +25,17 @@ class DefaultUserSessionServiceTest {
 
     @Test
     void testRegisterUser() {
-        userSessionService.secureHttpClient = [get : {url -> "[{\"username\":\"another\"}]"}] as SecureHttpClient
-        userSessionService.repository = [filter: { f -> [new User()]}] as Repository<User>
+        userSessionService.secureHttpClient = [get: { url -> "[{\"username\":\"another\"}]" }] as SecureHttpClient
+        userSessionService.repository = [filter: { f -> [new User()] }] as Repository<User>
 
         assert !userSessionService.registerUser(new RegistrationRequest(username: "another", password: "123456", email: "another@trevorism.com"))
     }
 
     @Test
     void testDoesUsernameExist() {
-        userSessionService.secureHttpClient = [get : {url -> "[{\"username\":\"another\"}]"}] as SecureHttpClient
+        userSessionService.secureHttpClient = [get: { url -> "[{\"username\":\"another\"}]" }] as SecureHttpClient
         userSessionService.repository = [filter: { f ->
-            if(f.simpleFilters[0].value == "another")
+            if (f.simpleFilters[0].value == "another")
                 return [new User(username: "another")]
             return null
         }] as Repository<User>
@@ -45,8 +45,8 @@ class DefaultUserSessionServiceTest {
 
     @Test
     void testValidate() {
-        userSessionService.secureHttpClient = [get : {url -> "[]"}] as SecureHttpClient
-        userSessionService.repository = [filter: { f -> null}] as Repository<User>
+        userSessionService.secureHttpClient = [get: { url -> "[]" }] as SecureHttpClient
+        userSessionService.repository = [filter: { f -> null }] as Repository<User>
 
         assert !userSessionService.validate(null)
         assert !userSessionService.validate(new RegistrationRequest())
@@ -59,36 +59,32 @@ class DefaultUserSessionServiceTest {
     @Test
     void testResetPassword() {
         userSessionService.forgotPasswordLinkRepository = [get: { x -> new ForgotPasswordLink() }, delete: { x -> new ForgotPasswordLink() }] as Repository
-        userSessionService.secureHttpClient = [post : {url, json -> "{}"}] as SecureHttpClient
+        userSessionService.secureHttpClient = [post: { url, json -> "{}" }] as SecureHttpClient
         userSessionService.resetPassword("reset")
         assert true
     }
 
-    @Test(expected = RuntimeException)
+    @Test
     void testGenerateForgotPasswordLink() {
         userSessionService.forgotPasswordLinkRepository = [get: { x -> new ForgotPasswordLink() }, delete: { x -> new ForgotPasswordLink() }] as Repository
-        userSessionService.httpClient = [get : { url, map ->
-            return ([getEntity: { new StringEntity("[]") }] as CloseableHttpResponse)
-        },
-                                         post: { url, content, map ->
-                                             return ([getEntity: { new StringEntity("token") }] as CloseableHttpResponse)
-                                         }] as HeadersJsonHttpClient
-        userSessionService.generateForgotPasswordLink(new ForgotPasswordRequest(email: "test@trevorism.com"))
+        userSessionService.httpClient = [get : { url, map -> return new HeadersHttpResponse("[]]") },
+                                         post: { url, content, map -> return new HeadersHttpResponse("token") }] as HttpClient
+        assertThrows(RuntimeException, () -> userSessionService.generateForgotPasswordLink(new ForgotPasswordRequest(email: "test@trevorism.com")))
     }
 
     @Test
     void testChangePassword() {
         userSessionService.httpClient = [post: { url, content, map ->
-            return ([getEntity: { new StringEntity("true") }] as CloseableHttpResponse)
-        }] as HeadersJsonHttpClient
+            return new HeadersHttpResponse("true")
+        }] as HttpClient
         assert userSessionService.changePassword(new ChangePasswordRequest(), "token")
     }
 
     @Test
     void testChangePasswordNotWorking() {
         userSessionService.httpClient = [post: { url, content, map ->
-            return ([getEntity: { new StringEntity("blah") }] as CloseableHttpResponse)
-        }] as HeadersJsonHttpClient
+            return new HeadersHttpResponse("blah")
+        }] as HttpClient
         assert !userSessionService.changePassword(new ChangePasswordRequest(), "token")
     }
 
